workers/proctor/requirements.txt
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
webrtcvad==2.0.10
pyannote.audio==3.1.1
torch==2.1.0
ultralytics==8.0.196
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
Pillow==10.0.1
ffmpeg-python==0.2.0 